move characterize_psf
