+++ scripts/beads_registration.py
@@ -3,7 +3,7 @@ import ants
 import numpy as np
 from pathlib import Path
 from iohub import open_ome_zarr
-from biahub.analysis.analyze_psf import detect_peaks
+from biahub.characterize_psf import detect_peaks
 from biahub.register import convert_transform_to_ants
 import napari
 from skimage.transform import EuclideanTransform, AffineTransform, warp
